+++ src/assets/icons/ResetIcon.tsx
@@ -1,0 +1,16 @@
+export const ResetIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+    <path d="M4.05 11a8 8 0 1 1 .5 4m-.5 5v-5h5" />
+  </svg>
+);